+++ src/ocean/util/app/LoggedCliApp.d
@@ -9,7 +9,7 @@
 
 *******************************************************************************/
 
-module ocean.util.app.LoggedCliApp;
+deprecated module ocean.util.app.LoggedCliApp;
 
 
 
@@ -103,6 +103,8 @@ import tango.util.log.Log;
 
 *******************************************************************************/
 
+deprecated("All applications should be migrated to use either DaemonApp or "
+           "CliApp")
 abstract class LoggedCliApp : ConfiguredCliApp, ILogExtExtension
 {